Fixed some bugs

+++ Int16.cpp
@@ -22,7 +22,7 @@ IOperand const * Int16::operator+( IOperand const &rhs ) const
 {
     eOperandType type = std::max(this->getType(), rhs.getType());
     if (type < eOperandType::FLOAT)
-        return Factory::instance().createOperand(type, std::to_string(std::stoi(this->_value) + std::stoi(rhs.getValue())));
+        return Factory::instance().createOperand(type, std::to_string(std::stoll(this->_value) + std::stoll(rhs.getValue())));
     else
         return Factory::instance().createOperand(type, std::to_string(std::stold(this->_value) + std::stold(rhs.getValue())));
 }
@@ -31,7 +31,7 @@ IOperand const * Int16::operator-( IOperand const &rhs ) const
 {
     eOperandType type = std::max(this->getType(), rhs.getType());
     if (type < eOperandType::FLOAT)
-        return Factory::instance().createOperand(type, std::to_string(std::stoi(this->_value) - std::stoi(rhs.getValue())));
+        return Factory::instance().createOperand(type, std::to_string(std::stoll(this->_value) - std::stoll(rhs.getValue())));
     else
         return Factory::instance().createOperand(type, std::to_string(std::stold(this->_value) - std::stold(rhs.getValue())));
 }
@@ -40,7 +40,7 @@ IOperand const * Int16::operator*( IOperand const &rhs ) const
 {
     eOperandType type = std::max(this->getType(), rhs.getType());
     if (type < eOperandType::FLOAT)
-        return Factory::instance().createOperand(type, std::to_string(std::stoi(this->_value) * std::stoi(rhs.getValue())));
+        return Factory::instance().createOperand(type, std::to_string(std::stoll(this->_value) * std::stoll(rhs.getValue())));
     else
         return Factory::instance().createOperand(type, std::to_string(std::stold(this->_value) * std::stold(rhs.getValue())));
 }
@@ -51,7 +51,7 @@ IOperand const * Int16::operator/( IOperand const &rhs ) const
         throw (VMException("EXCEPTION: Division by 0."));
     eOperandType type = std::max(this->getType(), rhs.getType());
     if (type < eOperandType::FLOAT)
-        return Factory::instance().createOperand(type, std::to_string(std::stoi(this->_value) / std::stoi(rhs.getValue())));
+        return Factory::instance().createOperand(type, std::to_string(std::stoll(this->_value) / std::stoll(rhs.getValue())));
     else
         return Factory::instance().createOperand(type, std::to_string(std::stold(this->_value) / std::stold(rhs.getValue())));
 }
@@ -62,7 +62,7 @@ IOperand const * Int16::operator%( IOperand const &rhs ) const
         throw (VMException("EXCEPTION: Modulo by 0."));
     eOperandType type = std::max(this->getType(), rhs.getType());
     if (type < eOperandType::FLOAT)
-        return Factory::instance().createOperand(type, std::to_string(std::stoi(this->_value) % std::stoi(rhs.getValue())));
+        return Factory::instance().createOperand(type, std::to_string(std::stoll(this->_value) % std::stoll(rhs.getValue())));
     else
         return Factory::instance().createOperand(type, std::to_string(static_cast<int64_t >(std::stold(this->_value)) % static_cast<int64_t >(std::stold(rhs.getValue()))));
 }
