+++ int8.cpp
@@ -34,7 +34,7 @@ IOperand const * Int8::operator+( IOperand const &rhs ) const
 {
     eOperandType type = std::max(this->getType(), rhs.getType());
     if (type < eOperandType::FLOAT)
-        return Factory::instance().createOperand(type, std::to_string(std::stoi(this->_value) + std::stoi(rhs.getValue())));
+        return Factory::instance().createOperand(type, std::to_string(std::stoll(this->_value) + std::stoll(rhs.getValue())));
     else
         return Factory::instance().createOperand(type, std::to_string(std::stold(this->_value) + std::stold(rhs.getValue())));
 }
@@ -43,7 +43,7 @@ IOperand const * Int8::operator-( IOperand const &rhs ) const
 {
     eOperandType type = std::max(this->getType(), rhs.getType());
     if (type < eOperandType::FLOAT)
-        return Factory::instance().createOperand(type, std::to_string(std::stoi(this->_value) - std::stoi(rhs.getValue())));
+        return Factory::instance().createOperand(type, std::to_string(std::stoll(this->_value) - std::stoll(rhs.getValue())));
     else
         return Factory::instance().createOperand(type, std::to_string(std::stold(this->_value) - std::stold(rhs.getValue())));
 }
@@ -52,7 +52,7 @@ IOperand const * Int8::operator*( IOperand const &rhs ) const
 {
     eOperandType type = std::max(this->getType(), rhs.getType());
     if (type < eOperandType::FLOAT)
-        return Factory::instance().createOperand(type, std::to_string(std::stoi(this->_value) * std::stoi(rhs.getValue())));
+        return Factory::instance().createOperand(type, std::to_string(std::stoll(this->_value) * std::stoll(rhs.getValue())));
     else
         return Factory::instance().createOperand(type, std::to_string(std::stold(this->_value) * std::stold(rhs.getValue())));
 }
@@ -63,7 +63,7 @@ IOperand const * Int8::operator/( IOperand const &rhs ) const
         throw (VMException("EXCEPTION: Division by 0."));
     eOperandType type = std::max(this->getType(), rhs.getType());
     if (type < eOperandType::FLOAT)
-        return Factory::instance().createOperand(type, std::to_string(std::stoi(this->_value) / std::stoi(rhs.getValue())));
+        return Factory::instance().createOperand(type, std::to_string(std::stoll(this->_value) / std::stoll(rhs.getValue())));
     else
         return Factory::instance().createOperand(type, std::to_string(std::stold(this->_value) / std::stold(rhs.getValue())));
 }
@@ -74,7 +74,7 @@ IOperand const * Int8::operator%( IOperand const &rhs ) const
         throw (VMException("EXCEPTION: Modulo by 0."));
     eOperandType type = std::max(this->getType(), rhs.getType());
     if (type < eOperandType::FLOAT)
-        return Factory::instance().createOperand(type, std::to_string(std::stoi(this->_value) % std::stoi(rhs.getValue())));
+        return Factory::instance().createOperand(type, std::to_string(std::stoll(this->_value) % std::stoll(rhs.getValue())));
     else
         return Factory::instance().createOperand(type, std::to_string(static_cast<int64_t >(std::stold(this->_value)) % static_cast<int64_t >(std::stold(rhs.getValue()))));
 }
